removed unnecessary handler proxy
events were updated in the latest refactor such that the event target is passed as the first argument to every handler. originally it was necessary to use a proxy function that would first extract the target out of the browser's native event object, then pass it to the actual handler.