Component and Container
Component - a base class for classes that are intended to represent view
components
Container - a Component which is intended to manage a collection of
Components and render them together in a div

These classes are not permanent; I am experimenting with them for now.
It may be that they add unnecessary complexity, in which case they will
either be removed completely or re-implemented.

Inheritance has also been refactored considerably to facilitate these
new classes. These changes, unlike the new classes, are permanent.